requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
argilla==1.5.1
async-timeout==4.0.2
attrs==22.2.0
backoff==2.2.1
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
commonmark==0.9.1
dataclasses-json==0.5.7
Deprecated==1.2.13
distlib==0.3.6
dnspython==2.3.0
et-xmlfile==1.1.0
fastapi==0.95.1
filelock==3.10.7
frozenlist==1.3.3
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
joblib==1.2.0
langchain==0.0.129
loguru==0.6.0
lxml==4.9.2
Markdown==3.4.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
monotonic==1.6
msg-parser==1.2.0
multidict==6.0.4
mypy-extensions==1.0.0
nltk==3.8.1
numpy==1.23.5
olefile==0.46
openai==0.27.2
openpyxl==3.1.2
packaging==23.0
pandas==1.5.3
Pillow==9.5.0
pinecone-client==2.2.1
platformdirs==3.2.0
pydantic==1.10.7
Pygments==2.14.0
pypandoc==1.11
python-dateutil==2.8.2
python-docx==0.8.11
python-magic==0.4.27
python-pptx==0.6.21
pytz==2023.3
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
rfc3986==1.5.0
rich==13.0.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
SQLAlchemy==1.4.47
starlette==0.26.1
tenacity==8.2.2
tqdm==4.65.0
typing-inspect==0.8.0
typing_extensions==4.5.0
unstructured==0.5.8
urllib3==1.26.15
uvicorn==0.21.1
virtualenv==20.21.0
win32-setctime==1.1.0
wrapt==1.14.1
XlsxWriter==3.0.9
yarl==1.8.2
